implemented repair elem

+++ Assets/Scripts/Damager.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class Damager : MonoBehaviour
 {
-    public int mDamage;
+    public float mDamage;
     // Start is called before the first frame update
     void Start()
     {
         
     }
 
-    public int GetDamage()
+    public float GetDamage()
     {
         return mDamage;
     }
